Consolidates styles to one file
