+++ components/Incrementer.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Text, TextInput, View, Button, StyleSheet } from 'react-native';
 import { connect } from 'react-redux';
+import styles from "../styles/styles.js";
+
 
 function Counter({ lastCompletedDay, latestMaxPullUps, dispatch }) {
   return (
@@ -11,15 +13,15 @@ function Counter({ lastCompletedDay, latestMaxPullUps, dispatch }) {
   );
 }
 
-const styles = StyleSheet.create({
-  input: {
-    height: 40,
-    margin: 12,
-    borderWidth: 1,
-    padding: 10,
-    color: 'green'
-  },
-});
+// const styles = StyleSheet.create({
+//   input: {
+//     height: 40,
+//     margin: 12,
+//     borderWidth: 1,
+//     padding: 10,
+//     color: 'green'
+//   },
+// });
 
 function mapStateToProps(state) {
   return {
