+++ components/Menu.js
@@ -1,4 +1,6 @@
 import { FlatList, View, Text, StyleSheet } from "react-native";
+import styles from "../styles/styles.js";
+
 // import { connect } from 'react-redux';
 
 const data = [
@@ -8,25 +10,25 @@ const data = [
 
 const render = (item) => {
   return (
-    <View style={styles.item}>
+    <View style={styles.menuItem}>
       <Text style={styles.title}>{item.item.title}</Text>
       {console.log(item)}
     </View>
   );
 };
 
-const styles = StyleSheet.create({
-  item: {
-    backgroundColor: '#f9c2ff',
-    padding: 20,
-    marginVertical: 8,
-    marginHorizontal: 16,
-  },
-  title: {
-    fontSize: 32,
+// const styles = StyleSheet.create({
+//   item: {
+//     backgroundColor: '#f9c2ff',
+//     padding: 20,
+//     marginVertical: 8,
+//     marginHorizontal: 16,
+//   },
+//   title: {
+//     fontSize: 32,
 
-  },
-});
+//   },
+// });
 
 export default function Menu() {
   return (
